Simple session management
`flask.session` uses a secret key to manage session data on top of a
user's cookies. The key, so long as it is unknown, allow for protection
against tampering of the user's cookies.

There is a lot of going on behind the scenes, which is nice. It is
somewhat interesting to call an object from the flask module and simply
accept it has the information of the incoming request. Example:
`flask.request`.