Update table view on order placement
Use HX-Trigger headers to cause the table to refresh itself on an order
placement. Ideally this would append the new row to the top of the
table. However it appears that `hx-swap-oob` does not pair well with
tables. The partial template's inner children were being appended while
the parent `<tr>` was being excluded.